+++ lib/rubyoctopus.rb
@@ -4,6 +4,7 @@ require_relative "rubyoctopus/version"
 require_relative "rubyoctopus/octopusclient"
 require_relative "rubyoctopus/model/environmentresource"
 require_relative "rubyoctopus/model/resource"
+require_relative "rubyoctopus/model/resourcecollection"
 
 module RubyOctopus
   class Error < StandardError; end