+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+- Add ResourceCollection model for paging
+- Add OctopusClient list and paginate methods
 
 ## [0.1.0] - 2022-01-26